+++ contentful_management/resource_builder.py
@@ -1,15 +1,15 @@
-# Classes imported here are meant to be used via globals() on build
-from .entry import Entry  # noqa: F401
-from .asset import Asset  # noqa: F401
-from .space import Space  # noqa: F401
-from .content_type import ContentType  # noqa: F401
-from .webhook import Webhook  # noqa: F401
-from .locale import Locale  # noqa: F401
-from .role import Role  # noqa: F401
-from .editor_interface import EditorInterface  # noqa: F401
-from .api_key import ApiKey  # noqa: F401
-from .snapshot import Snapshot  # noqa: F401
-from .upload import Upload  # noqa: F401
+from .entry import Entry
+from .asset import Asset
+from .space import Space
+from .content_type import ContentType
+from .webhook import Webhook
+from .locale import Locale
+from .role import Role
+from .editor_interface import EditorInterface
+from .api_key import ApiKey
+from .snapshot import Snapshot
+from .upload import Upload
+from .array import Array
 
 
 """
@@ -39,30 +39,27 @@ class ResourceBuilder(object):
         return self._build_item(self.json)
 
     def _build_array(self):
-        return [self._build_item(item) for item in self.json['items']]
+        return Array(self.json, [self._build_item(item) for item in self.json['items']])
 
     def _build_item(self, item):
-        buildables = [
-            'Entry',
-            'Asset',
-            'ContentType',
-            'Space',
-            'ApiKey',
-            'Locale',
-            'EditorInterface',
-            'Webhook',
-            'Role',
-            'Snapshot',
-            'Upload'
-        ]
+        buildables = {
+            'Entry': Entry,
+            'Asset': Asset,
+            'ContentType': ContentType,
+            'Space': Space,
+            'ApiKey': ApiKey,
+            'Locale': Locale,
+            'EditorInterface': EditorInterface,
+            'WebhookDefinition': Webhook,
+            'Role': Role,
+            'Snapshot': Snapshot,
+            'Upload': Upload
+        }
 
         item_type = item['sys']['type']
 
-        if item_type == 'WebhookDefinition':
-            item_type = 'Webhook'
-
         if item_type in buildables:
-            return globals()[item_type](
+            return buildables[item_type](
                 item,
                 default_locale=self.default_locale,
                 client=self.client